added fail-safe mechanisms, and Mac OS build

+++ distribution/build_dev.csh
@@ -1,30 +1,48 @@
 #!/bin/tcsh -f
 
-
 set ECHO=
+#set echo=1
+
+set HOSTNAME=`hostname -s`
 
 # Set up directories.
 ######################################################################
 if(! $?BUILD_DIR) then
-  setenv BUILD_DIR /space/birn/50/freesurfer/build/`hostname -s`
+  setenv BUILD_DIR /space/freesurfer/build/$HOSTNAME
 endif
 
-setenv QTDIR /space/birn/50/freesurfer/build/`hostname -s`/qt
-setenv LD_LIBRARY_PATH ${QTDIR}/lib
-setenv DYLD_LIBRARY_PATH ${QTDIR}/lib
+setenv QTDIR /space/freesurfer/build/$HOSTNAME/qt
+setenv GLUT_DYLIB_DIR ""
 
-set SCRIPT_DIR=/space/birn/50/freesurfer/build/scripts
-set LOG_DIR=/space/birn/50/freesurfer/build/logs
+set SCRIPT_DIR=/space/freesurfer/build/scripts
+set LOG_DIR=/space/freesurfer/build/logs
+
+# automount doesnt work on Mac OS X Tiger, 
+# so prefacing paths with /autofs is a kluge.
+# also, need /sw/bin to get latex and dvips
+if ("$HOSTNAME" == "storm") then
+    setenv BUILD_DIR /autofs${BUILD_DIR}
+    setenv QTDIR /autofs${QTDIR}
+    set SCRIPT_DIR=/autofs${SCRIPT_DIR}
+    set LOG_DIR=/autofs${LOG_DIR}
+    set GLUT_DYLIB_DIR=/autofs/space/freesurfer/build/storm/tiffjpegglut/lib
+    setenv PATH "/sw/bin":"$PATH"
+    rehash
+endif
+
 set DEV_DIR=${BUILD_DIR}/trunk/dev
 set DEV_DEST_DIR=/usr/local/freesurfer/dev
-
 set FAILED_FILE=${BUILD_DIR}/dev-FAILED
+
+setenv LD_LIBRARY_PATH "${QTDIR}/lib":"${GLUT_DYLIB_DIR}"
+setenv DYLD_LIBRARY_PATH "${QTDIR}/lib":"${GLUT_DYLIB_DIR}"
 
 # Output file
 ######################################################################
-set MAIL_LIST=(kteich@nmr.mgh.harvard.edu)
-set OUTPUTF=/tmp/build_log-dev-`hostname -s`.txt
-echo "`hostname -s` build" >& $OUTPUTF
+set MAIL_LIST=(kteich@nmr.mgh.harvard.edu nicks@nmr.mgh.harvard.edu)
+set OUTPUTF=$LOG_DIR/build_log-dev-$HOSTNAME.txt
+echo "$HOSTNAME build" >& $OUTPUTF
+$ECHO chmod g+w $OUTPUTF
 set BEGIN_TIME=`date`
 echo $BEGIN_TIME >>& $OUTPUTF
 
@@ -36,13 +54,12 @@ set OS=`uname -s`
 ######################################################################
 if(! -d $SCRIPT_DIR) then 
   echo "$SCRIPT_DIR doesn't exist" >>& $OUTPUTF
-  $ECHO mail -s "`hostname -s` dev build FAILED - sanity" $MAIL_LIST < $OUTPUTF
+  $ECHO mail -s "$HOSTNAME dev build FAILED - sanity" $MAIL_LIST < $OUTPUTF
   exit 1  
 endif
-
 if(! -d $DEV_DIR) then 
   echo "$DEV_DIR doesn't exist" >>& $OUTPUTF
-  $ECHO mail -s "`hostname -s` dev build FAILED - sanity" $MAIL_LIST < $OUTPUTF
+  $ECHO mail -s "$HOSTNAME dev build FAILED - sanity" $MAIL_LIST < $OUTPUTF
   exit 1  
 endif
 
@@ -91,30 +108,38 @@ echo "Updating dev" >>& $OUTPUTF
 echo "" >>& $OUTPUTF
 $ECHO echo "CMD: cd $DEV_DIR" >>& $OUTPUTF
 $ECHO cd $DEV_DIR >>& $OUTPUTF
-$ECHO echo "CMD: cvs update -d \>\& /tmp/update-output" >>& $OUTPUTF
-$ECHO cvs update -d >& /tmp/update-output
+$ECHO echo "CMD: cvs update -d \>\& $LOG_DIR/update-output-$HOSTNAME" >>& $OUTPUTF
+$ECHO cvs update -d >& $LOG_DIR/update-output-$HOSTNAME
 
-$ECHO echo "CMD: grep -e ^\[UP\]\  /tmp/update-output" >>& $OUTPUTF
-$ECHO grep -e ^\[UP\]\  /tmp/update-output >& /dev/null
+$ECHO echo "CMD: grep -e ^\[UP\]\  $LOG_DIR/update-output-$HOSTNAME" >>& $OUTPUTF
+$ECHO grep -e ^\[UP\]\  $LOG_DIR/update-output-$HOSTNAME >& /dev/null
 if ($status != 0 && ! -e ${FAILED_FILE} ) then
   echo "Nothing changed in repository, SKIPPED building" >>& $OUTPUTF
   goto done
 endif
 
-$ECHO echo "CMD: cat /tmp/update-output \>\>\& $OUTPUTF" >>& $OUTPUTF
-$ECHO cat /tmp/update-output >>& $OUTPUTF
-$ECHO echo "CMD: rm -f /tmp/update-output" >>& $OUTPUTF
-$ECHO rm -f /tmp/update-output
+# assume failure (file removed only after successful build)
+$ECHO touch ${FAILED_FILE}
+$ECHO chmod g+w ${FAILED_FILE}
 
+$ECHO echo "CMD: cat $LOG_DIR/update-output-$HOSTNAME \>\>\& $OUTPUTF" >>& $OUTPUTF
+$ECHO cat $LOG_DIR/update-output-$HOSTNAME >>& $OUTPUTF
+$ECHO echo "CMD: rm -f $LOG_DIR/update-output-$HOSTNAME" >>& $OUTPUTF
+$ECHO rm -f $LOG_DIR/update-output-$HOSTNAME
 
 
 echo "##########################################################" >>& $OUTPUTF
 echo "Freshening Makefiles" >>& $OUTPUTF
 echo "" >>& $OUTPUTF
 $ECHO echo "CMD: make distclean" >>& $OUTPUTF
-$ECHO make distclean >>& $OUTPUTF
+if (-e Makefile) make distclean >>& $OUTPUTF
 $ECHO echo "CMD: rm -rf autom4te.cache" >>& $OUTPUTF
-$ECHO rm -rf autom4te.cache >>& $OUTPUTF
+if (-e autom4te.cache) rm -rf autom4te.cache >>& $OUTPUTF
+$ECHO echo "CMD: libtoolize --force" >>& $OUTPUTF
+if ( "`uname -s`" == "Linux") libtoolize --force >>& $OUTPUTF
+if ( "`uname -s`" == "Darwin") glibtoolize --force >>& $OUTPUTF
+$ECHO echo "CMD: autoreconf --force" >>& $OUTPUTF
+$ECHO autoreconf --force >>& $OUTPUTF
 $ECHO echo "CMD: aclocal" >>& $OUTPUTF
 $ECHO aclocal >>& $OUTPUTF
 $ECHO echo "CMD: autoconf" >>& $OUTPUTF
@@ -128,27 +153,57 @@ if ($status != 0) then
   echo "config.log" >>& $OUTPUTF
   echo "" >>& $OUTPUTF
   $ECHO cat ${DEV_DIR}/config.log >>& $OUTPUTF
-  $ECHO mail -s "`hostname -s` dev build FAILED after dev configure" $MAIL_LIST < $OUTPUTF
+  $ECHO mail -s "$HOSTNAME dev build FAILED after dev configure" $MAIL_LIST < $OUTPUTF
   $ECHO touch ${FAILED_FILE}
+  $ECHO chmod g+w ${FAILED_FILE}
+  # set group write bit on files changed by make tools:
+  $ECHO echo "CMD: chmod -R g+rw ${DEV_DIR}" >>& $OUTPUTF
+  $ECHO chmod -R g+rw ${DEV_DIR} >>& $OUTPUTF
   exit 1  
 endif
 
 echo "##########################################################" >>& $OUTPUTF
 echo "Building dev" >>& $OUTPUTF
 echo "" >>& $OUTPUTF
+# make
+$ECHO echo "CMD: make" >>& $OUTPUTF
+$ECHO make >>& $OUTPUTF
+if ($status != 0) then
+  # note: /usr/local/freesurfer/dev/bin/ dirs have not been modified (make install does that)
+  $ECHO mail -s "$HOSTNAME dev build (make) FAILED" $MAIL_LIST < $OUTPUTF
+  $ECHO touch ${FAILED_FILE}
+  $ECHO chmod g+w ${FAILED_FILE}
+  # set group write bit on files changed by make tools:
+  $ECHO echo "CMD: chmod -R g+rw ${DEV_DIR}" >>& $OUTPUTF
+  $ECHO chmod -R g+rw ${DEV_DIR} >>& $OUTPUTF
+  exit 1  
+endif
+
 # Shift bin to bin-old and bin-old to bin-old-old to keep around old versions.
 $ECHO echo "CMD: rm -rf ${DEV_DEST_DIR}/bin-old-old" >>& $OUTPUTF
-$ECHO rm -rf ${DEV_DEST_DIR}/bin-old-old >>& $OUTPUTF
+if (-e ${DEV_DEST_DIR}/bin-old-old) rm -rf ${DEV_DEST_DIR}/bin-old-old >>& $OUTPUTF
 $ECHO echo "CMD: mv ${DEV_DEST_DIR}/bin-old ${DEV_DEST_DIR}/bin-old-old" >>& $OUTPUTF
-$ECHO mv ${DEV_DEST_DIR}/bin-old ${DEV_DEST_DIR}/bin-old-old >>& $OUTPUTF
+if (-e ${DEV_DEST_DIR}/bin-old) mv ${DEV_DEST_DIR}/bin-old ${DEV_DEST_DIR}/bin-old-old >>& $OUTPUTF
 $ECHO echo "CMD: mv ${DEV_DEST_DIR}/bin ${DEV_DEST_DIR}/bin-old" >>& $OUTPUTF
 $ECHO mv ${DEV_DEST_DIR}/bin ${DEV_DEST_DIR}/bin-old >>& $OUTPUTF
 
+# make install
 $ECHO echo "CMD: make install" >>& $OUTPUTF
 $ECHO make install >>& $OUTPUTF
 if ($status != 0) then
-  $ECHO mail -s "`hostname -s` dev build FAILED after building dev" $MAIL_LIST < $OUTPUTF
+  $ECHO mail -s "$HOSTNAME dev build (make install) FAILED" $MAIL_LIST < $OUTPUTF
   $ECHO touch ${FAILED_FILE}
+  $ECHO chmod g+w ${FAILED_FILE}
+  # restore prior (possibly working) bin/ dirs
+  $ECHO echo "CMD: mv ${DEV_DEST_DIR}/bin ${DEV_DEST_DIR}/bin-failed" >>& $OUTPUTF
+  $ECHO mv ${DEV_DEST_DIR}/bin ${DEV_DEST_DIR}/bin-failed >>& $OUTPUTF
+  $ECHO echo "CMD: mv ${DEV_DEST_DIR}/bin-old ${DEV_DEST_DIR}/bin" >>& $OUTPUTF
+  $ECHO mv ${DEV_DEST_DIR}/bin-old ${DEV_DEST_DIR}/bin >>& $OUTPUTF
+  $ECHO echo "CMD: mv ${DEV_DEST_DIR}/bin-old-old ${DEV_DEST_DIR}/bin-old" >>& $OUTPUTF
+  $ECHO mv ${DEV_DEST_DIR}/bin-old-old ${DEV_DEST_DIR}/bin-old >>& $OUTPUTF
+  # set group write bit on files changed by make tools:
+  $ECHO echo "CMD: chmod -R g+rw ${DEV_DIR}" >>& $OUTPUTF
+  $ECHO chmod -R g+rw ${DEV_DIR} >>& $OUTPUTF
   exit 1  
 endif
 
@@ -157,7 +212,12 @@ echo "Setting permissions" >>& $OUTPUTF
 echo "" >>& $OUTPUTF
 $ECHO echo "CMD: chmod -R g+rw ${DEV_DEST_DIR}" >>& $OUTPUTF
 $ECHO chmod -R g+rw ${DEV_DEST_DIR} >>& $OUTPUTF
+$ECHO echo "CMD: chmod -R g+rw ${DEV_DIR}" >>& $OUTPUTF
+$ECHO chmod -R g+rw ${DEV_DIR} >>& $OUTPUTF
+$ECHO echo "CMD: chmod -R g+rw ${LOG_DIR}" >>& $OUTPUTF
+$ECHO chmod -R g+rw ${LOG_DIR} >>& $OUTPUTF
 
+# Success, so remove fail indicator:
 $ECHO rm -rf ${FAILED_FILE}
 
 done:
@@ -171,13 +231,13 @@ echo $END_TIME >>& $OUTPUTF
 ######################################################################
 
 # Move log file to stamped version.
-$ECHO mv $OUTPUTF ${LOG_DIR}/build_log-dev-`hostname -s`-$TIME_STAMP.txt
-$ECHO gzip -f ${LOG_DIR}/build_log-dev-`hostname -s`-$TIME_STAMP.txt
+$ECHO mv $OUTPUTF ${LOG_DIR}/build_log-dev-$HOSTNAME-$TIME_STAMP.txt
+$ECHO gzip -f ${LOG_DIR}/build_log-dev-$HOSTNAME-$TIME_STAMP.txt
 
 # Send email.
-echo "Begin ${BEGIN_TIME}, end ${END_TIME}" >& /tmp/message.txt
-$ECHO mail -s "`hostname -s` dev build is awesome." $MAIL_LIST < /tmp/message.txt
-$ECHO rm /tmp/message.txt
+echo "Begin ${BEGIN_TIME}, end ${END_TIME}" >& $LOG_DIR/message-$HOSTNAME.txt
+$ECHO mail -s "$HOSTNAME dev build is wicked awesome." $MAIL_LIST < $LOG_DIR/message-$HOSTNAME.txt
+$ECHO rm $LOG_DIR/message-$HOSTNAME.txt
 
 # Soon:
 # $ECHO make install ${x8664BUILDCXXFLAGS} bindir=\\\$\\\{prefix\\\}/bin/x86-64
